Add files via upload
A 500px wide and 50px tall color strip that will determine the hue of the gradient, based on the cursor position. Below it is a 500 x 500 gradient showing the different saturation and brightness values for that hue. On top is the brightest and gets darker as you go down, while the leftmost is the most saturated and gets less saturated towards the right.